requirements.txt
SQLAlchemy==2.0.35
cachetools==5.3.3
fastapi==0.115.4
lxml==5.3.0
pydantic==2.9.2
python-dotenv==1.0.1
rapidfuzz==3.9.6
requests==2.32.3
uvicorn[standard]==0.30.6
